requirements.txt
# Core AI/ML libraries
openai>=1.0.0
# Pin httpx below 0.28 due to breaking change removing 'proxies' kwarg
# OpenAI SDK (some versions) passes 'proxies' to httpx.Client; httpx>=0.28 breaks this
httpx<0.28
elevenlabs>=0.2.0

# Voice processing
SpeechRecognition>=3.10.0
openai-whisper>=20231117

# Audio handling
pygame>=2.5.0

# System monitoring
psutil>=5.9.0

# Environment management
python-dotenv>=1.0.0

# GUI (for future Qt integration)
# PyQt6>=6.4.0

# File processing (for future features)
PyMuPDF>=1.23.0
python-docx>=0.8.11
pandas>=2.0.0

# Optional: Additional audio libraries
# pydub>=0.25.1
# soundfile>=0.12.1